確認画面から戻るボタン追加 その他、コメントや軽微な修正

+++ fuel/app/classes/controller/staff.php
@@ -37,10 +37,17 @@ class Controller_Staff extends Controller_Hybrid
 
         $view->set_global('title', "Staff Add");
 
-        // 定義呼出
         Config::load('staff_master', true);
         $view->set_global('department_arr', Config::get('staff_master.department'));
         $view->set_global('gender_arr', Config::get('staff_master.gender'));
+
+        if (Session::get_flash('action')) {
+            foreach ($this->staffs as $staff) {
+                // セッションからフラッシュ変数を取出
+                $staffs[$staff] = Session::get_flash($staff);
+            }
+            $view->set_global('staff', $staffs);
+        }
 
         // 入力チェック呼出
         $val = Model_Staff::validate('add');
@@ -49,7 +56,7 @@ class Controller_Staff extends Controller_Hybrid
         // POSTされた各データをフラッシュセッションに保存
         if (Input::method() == 'POST') {
 
-            // セッションにフラッシュ変数をセット
+            // セッションからフラッシュ変数を取出
             foreach ($this->staffs as $staff) {
                 Session::set_flash($staff, Input::post($staff));
             }
@@ -66,64 +73,6 @@ class Controller_Staff extends Controller_Hybrid
 
         $this->template->title = 'Staff &raquo; Add';
         $this->template->content = $view;
-    }
-
-    /**
-     * 確認画面
-     */
-    public function action_conf()
-    {
-        $view = \View::forge('staff/conf');
-
-        $view->set('title', "Staff Conf");
-
-        // 定義呼出
-        Config::load('staff_master', true);
-        $view->set('department_arr', Config::get('staff_master.department'));
-        $view->set('gender_arr', Config::get('staff_master.gender'));
-
-        foreach ($this->staffs as $staff) {
-            // セッションからフラッシュ変数を取出
-           $view->set($staff, Session::get_flash($staff));
-            // セッション変数を次のリクエストを維持
-            Session::keep_flash($staff);
-        }
-
-        $view->set('action', Session::get_flash('action'));
-
-        $this->template->title = 'Staff &raquo; Conf';
-        $this->template->content = $view;
-    }
-
-    /**
-     * DBに保存
-     */
-    public function action_insert()
-    {
-        if (Security::check_token()) {
-            // postされたデータを取り出す
-            $val = [
-                'staff_no'        => Input::post('staff_no')
-                , 'name'          => Input::post('name')
-                , 'department'    => Input::post('department')
-                , 'gender'        => Input::post('gender')
-            ];
-
-            // トランザクション
-            try {
-                DB::start_transaction();
-
-                Model_Staff::staff_insert_query($val);
-
-                DB::commit_transaction();
-
-                Response::redirect('staff');
-            } catch (\Exception $e) {
-                DB::rollback_transaction();
-
-                throw $e;
-            }
-        }
     }
 
     /**
@@ -168,7 +117,15 @@ class Controller_Staff extends Controller_Hybrid
         $view->set_global('gender_arr', Config::get('staff_master.gender'));
 
 
-        $view->set_global('staff', Model_Staff::staff_detail_query($id));
+        if (Session::get_flash('action')) {
+            foreach ($this->staffs as $staff) {
+                // セッションからフラッシュ変数を取出
+                $staffs[$staff] = Session::get_flash($staff);
+            }
+            $view->set_global('staff', $staffs);
+        }else{
+            $view->set_global('staff', Model_Staff::staff_detail_query($id));
+        }
 
         // 入力チェック呼出
         $val = Model_Staff::validate('edit', $id);
@@ -200,11 +157,39 @@ class Controller_Staff extends Controller_Hybrid
     }
 
     /**
-     * DB更新
+     * 確認画面
      */
-    public function action_update()
+    public function action_conf()
+    {
+        $view = \View::forge('staff/conf');
+
+        $view->set('title', "Staff Conf");
+
+        // 定義呼出
+        Config::load('staff_master', true);
+        $view->set('department_arr', Config::get('staff_master.department'));
+        $view->set('gender_arr', Config::get('staff_master.gender'));
+
+        foreach ($this->staffs as $staff) {
+            // セッションからフラッシュ変数を取出
+           $view->set($staff, Session::get_flash($staff));
+            // セッション変数を次のリクエストを維持
+            Session::keep_flash($staff);
+        }
+
+        $view->set('action', Session::get_flash('action'));
+
+        $this->template->title = 'Staff &raquo; Conf';
+        $this->template->content = $view;
+    }
+
+    /**
+     * DBに保存
+     */
+    public function post_insert()
     {
         if (Security::check_token()) {
+            // postされたデータを取り出す
             $val = [
                 'staff_no'        => Input::post('staff_no')
                 , 'name'          => Input::post('name')
@@ -212,9 +197,39 @@ class Controller_Staff extends Controller_Hybrid
                 , 'gender'        => Input::post('gender')
             ];
 
+            // トランザクション
+            try {
+                DB::start_transaction();
+
+                Model_Staff::staff_insert_query($val);
+
+                DB::commit_transaction();
+
+                Response::redirect('staff');
+            } catch (\Exception $e) {
+                DB::rollback_transaction();
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
+     * DB更新
+     */
+    public function post_update()
+    {
+        if (Security::check_token()) {
 
             $id = Session::get('id');
             Session::delete('id');
+
+            $val = [
+                'staff_no'        => Input::post('staff_no')
+                , 'name'          => Input::post('name')
+                , 'department'    => Input::post('department')
+                , 'gender'        => Input::post('gender')
+            ];
 
             // トランザクション
             try {
@@ -234,11 +249,42 @@ class Controller_Staff extends Controller_Hybrid
     }
 
     /**
+     * 確認画面から戻る
+     */
+    public function post_back()
+    {
+        if (Input::method() == 'POST' && Security::check_token()) {
+
+            // IDのセッションがあれば編集画面へ戻る、無ければ新規の方に戻る
+            if ( Session::get('id') )
+            {
+                $id = Session::get('id');
+                Session::delete('id');
+
+                $url = 'edit/'.$id;
+            }else{
+                $url = 'add';
+            }
+
+            // セッション期限延期
+            foreach ($this->staffs as $staff)
+            {
+                Session::keep_flash($staff);
+            }
+
+            // 戻るからEditページ等を
+            Session::set_flash('action', 'back');
+
+            //入力画面にリダイレクト
+            Response::redirect('staff/'.$url);
+        }
+    }
+
+    /**
      * DB削除
      */
-    public function post_destory()
+    public function post_destory($id)
     {
-        $id = Input::post('id');
         if (Input::method() == 'POST' && Security::check_token()) {
             // トランザクション
             try {
